// dashboard/src/client/models/node-error-event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ResultError } from './result-error';

export interface NodeErrorEvent {
    'kind': NodeErrorEventKindEnum;
    'runId': string;
    'nodeId': string;
    'error': ResultError;
}

export const NodeErrorEventKindEnum = {
    NodeError: 'node.error'
} as const;

export type NodeErrorEventKindEnum = typeof NodeErrorEventKindEnum[keyof typeof NodeErrorEventKindEnum];